refactor: split rewrite helpers to its own class
This is motivated by the support for gRPC, as I will need to make this
helper work with that transport too.
