+++ gcs/holder.py
@@ -232,40 +232,3 @@ class DataHolder(types.SimpleNamespace):
             response.headers["Range"] = "bytes=0-%d" % (len(self.media) - 1)
         response.status_code = 308
         return response
-
-    # === REWRITE === #
-
-    # TODO(#22) - this should be in a separate class
-    @classmethod
-    def init_rewrite_rest(
-        cls, request, src_bucket_name, src_object_name, dst_bucket_name, dst_object_name
-    ):
-        fake_request = testbench.common.FakeRequest(
-            args=request.args.to_dict(),
-            headers={
-                key.lower(): value
-                for key, value in request.headers.items()
-                if key.lower().startswith("x-")
-            },
-            data=request.data,
-        )
-        max_bytes_rewritten_per_call = min(
-            int(fake_request.args.get("maxBytesRewrittenPerCall", 1024 * 1024)),
-            1024 * 1024,
-        )
-        token = hashlib.sha256(
-            (
-                "%s/o/%s/rewriteTo/b/%s/o/%s"
-                % (src_bucket_name, src_object_name, dst_bucket_name, dst_object_name)
-            ).encode("utf-8")
-        ).hexdigest()
-        return cls(
-            request=fake_request,
-            src_bucket_name=src_bucket_name,
-            src_object_name=src_object_name,
-            dst_bucket_name=dst_bucket_name,
-            dst_object_name=dst_object_name,
-            token=token,
-            media=b"",
-            max_bytes_rewritten_per_call=max_bytes_rewritten_per_call,
-        )
